Add game hanlder
